requirements.txt
flask
pyzbar
pillow
firebase-admin
